before 2 day a finish

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: mayache- <mayache-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/01 10:47:47 by mayache-          #+#    #+#             */
-/*   Updated: 2023/02/23 21:29:06 by mayache-         ###   ########.fr       */
+/*   Updated: 2023/04/17 21:23:57 by mayache-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long long	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int		i;
 	int		sign;
